6A-ii: delete the 18 analytic forks; phase6_scaffolding -> phase6_wired
With the recursion routed through HamiltonianOp (6A-i), the _Device / _Graphene_NNN
analytic forks are dead -- every golden uses materialised .CSR operators, so the
sparse adapter is the only live HamiltonianOp.

- Delete the 18 fork files (chebyshev_{moments,solver,vectors,moments1D,momentsTD}_
  {Device,Graphene_NNN}.{cpp,hpp}, device-postProcess_Graphene_NNN, and the 4 fork
  inline drivers).
- Drop the 9 fork sources from KPM_LIB_SOURCES and the 5 fork executables from
  LINQT_APPS. The Graphene_NNN / Device model classes (Device/*.cpp, Graphene_NNN.hpp)
  stay -- they back the kept device-postProcess driver and the future matrix-free
  _hr.dat backlog.
- Replace phase6_scaffolding with phase6_wired: compile-time asserts that
  Moments::op() returns HamiltonianOp&, metric() returns MetricPolicy&, and
  SparseHamiltonianOp implements the interface; placeholders still throw.

tests: ctest all green (24/24); exact-trace goldens byte-identical; tree builds with
the 18 forks gone and no dangling LINQT_APPS / KPM_LIB_SOURCES entries.
