--- src/inline_compute-kpm-MeanSquareDisplacement_DeviceSim.cpp
@@ -1,141 +1,0 @@
-// C & C++ libraries
-#include <iostream> /* for std::cout mostly */
-#include <string>   /* for std::string class */
-#include <fstream>  /* for std::ofstream and std::ifstream functions classes*/
-#include <stdlib.h>
-#include <chrono>
-
-
-#include "kpm_noneqop.hpp" //Message functions
-#include "chebyshev_moments.hpp"
-#include "sparse_matrix.hpp"
-#include "quantum_states.hpp"
-#include "chebyshev_solver.hpp"
-
-
-namespace time_evolution
-{
-	void printHelpMessage();
-	void printWelcomeMessage();
-}
-
-
-int main(int argc, char *argv[])
-{
-	if ( !(argc == 4 || argc == 5 ) )
-	{
-		time_evolution::printHelpMessage();
-		return 0;
-	}
-	else
-		time_evolution::printWelcomeMessage();
-	
-	const std::string
-		LABEL  = argv[1],
-		S_NMOM = argv[2],
-		S_NTIME= argv[3],
-		S_TMAX = argv[4];
-
-	const int numMoms = atoi(S_NMOM.c_str() );
-	const int numTimes= atoi(S_NTIME.c_str() );
-	const double tmax = stod(S_TMAX );
-
-	chebyshev::MomentsTD chebMoms(numMoms, numTimes); //load number of moments
-
-	
-	SparseMatrixType OP[3];
-	OP[0].SetID("HAM");
-	OP[1].SetID("VX");
-	OP[2].SetID("VY");
-	
-	
-	// Build the operators from Files
-	SparseMatrixBuilder builder;
-	std::array<double,2> spectral_bounds;	
-	for (int i = 0; i < 3; i++)
-	if( OP[i].isIdentity() )
-			std::cout<<"The operator "<<OP[i].ID()<<" is treated as the identity"<<std::endl;
-	else
-	{
-			std::string input = "operators/" + LABEL + "." + OP[i].ID() + ".CSR";
-			builder.setSparseMatrix(&OP[i]);
-			builder.BuildOPFromCSRFile(input);
-	
-			if( i == 0 ) //is hamiltonian
-			//Obtain automatically the energy bounds
-		 	spectral_bounds = chebyshev::utility::SpectralBounds(OP[0]);
-	};
-	//CONFIGURE THE CHEBYSHEV MOMENTS
-	chebMoms.SystemLabel(LABEL);
-	chebMoms.BandWidth ( (spectral_bounds[1]-spectral_bounds[0])*1.0);
-	chebMoms.BandCenter( (spectral_bounds[1]+spectral_bounds[0])*0.5);
-	chebMoms.TimeDiff( tmax/(numTimes) );
-	chebMoms.SetAndRescaleHamiltonian(OP[0]);
-
-	chebMoms.SetAndRescaleVX_VY(OP[1],OP[2]);
-	//Rescale the velocity
-	//OP[1].Rescale(chebMoms.ScaleFactor(),0.0);
-	chebMoms.Print();
-
-	//Compute the chebyshev expansion table
-	qstates::generator gen;
-	if( argc == 6)
-		gen  = qstates::LoadStateFile(argv[5]);
-
-	chebyshev::MeanSquareDisplacement_DeviceSim(chebMoms, numMoms, tmax/(numTimes-1), tmax, gen);
-
-
-
-		//Post-processing data;
-	const double a     = spectral_bounds[1] - spectral_bounds[0];
-	const double b     = (spectral_bounds[1] + spectral_bounds[0]) * 0.5;
-	const double tstep = tmax / double(numTimes);
-
-	//Standard values for postprocessing
-	const double Emin  = -1.0;
-	const double Emax  =  1.0;
-	const double dE    =  0.001;
-	const double eta   =  0.00026;
-
-	std::ofstream pp("postProcess.dat");
-	if (!pp) throw std::runtime_error("Cannot open postProcess.dat");
-
-	pp << "a         " << a         << "\n"
-	   << "b         " << b         << "\n"
-	   << "nMom_DOS  " << numMoms  << "\n"
-	   << "size      " << 1.0      << "\n"
-	   << "numTimes  " << numTimes  << "\n"
-
-	   << "tmax      " << tmax      << "\n"
-	   << "tstep     " << tstep     << "\n"
-
-	  //These may be altered afterwards
-	   << "Emin      " << Emin      << "\n"
-	   << "Emax      " << Emax      << "\n"
-	   << "dE        " << dE        << "\n"
-	   << "eta       " << eta       << "\n";
-
-	pp.close();
-
-
-
-
-	std::cout<<"End of program"<<std::endl;
-	return 0;
-}
-
-void time_evolution::printHelpMessage()
-	{
-		std::cout << "The program should be called with the following options: Label numMom numTimeSteps MaxTime (optional) num_states_file" << std::endl
-				  << std::endl;
-		std::cout << "Label will be used to look for Label.Ham, Label.VX and Label.VY " << std::endl;
-		std::cout << "numMom will be used to set the number of moments in the chebyshev table" << std::endl;
-		std::cout << "numTimeSteps  will be used to set the number of timesteps in the chebyshev table" << std::endl;
-		std::cout << "TimeMax  will be set the maximum time where the correlation will be evaluted " << std::endl;
-	};
-
-	inline
-void time_evolution::printWelcomeMessage()
-	{
-		std::cout << "WELCOME: This program will compute a table needed for expanding the correlation function in Chebyshev polynomialms" << std::endl;
-	};
