--- src/inline_compute-kpm-TimeEvOp_Device.cpp
@@ -1,148 +1,0 @@
-// C & C++ libraries
-#include <iostream> /* for std::cout mostly */
-#include <string>   /* for std::string class */
-#include <fstream>  /* for std::ofstream and std::ifstream functions classes*/
-#include <stdlib.h>
-#include <chrono>
-
-
-#include "kpm_noneqop.hpp" //Message functions
-#include "chebyshev_moments.hpp"
-
-#include "chebyshev_moments_Device.hpp"
-#include "sparse_matrix.hpp"
-#include "quantum_states.hpp"
-#include "chebyshev_solver_Device.hpp"
-#include "chebyshev_solver.hpp"
-#include "Device/Device.hpp"
-#include "Device/Graphene_KaneMele.hpp"
-
-namespace time_evolution
-{
-	void printHelpMessage();
-	void printWelcomeMessage();
-}
-
-
-int main(int argc, char *argv[])
-{
-	if ( !(argc == 6 || argc == 7 ) )
-	{
-		time_evolution::printHelpMessage();
-		return 0;
-	}
-	else
-		time_evolution::printWelcomeMessage();
-	
-	const std::string
-		LABEL  = argv[1],
-		S_OP   = argv[2],
-		S_NMOM = argv[3],
-		S_NTIME= argv[4],
-		S_TMAX = argv[5];
-
-	const int numMoms = atoi(S_NMOM.c_str() );
-	const int numTimes= atoi(S_NTIME.c_str() );
-	const double tmax = stod(S_TMAX );
-
-
-
-
-	bool is_k_space = true;
-        r_type stgr_str=0.0, m_str=0.0, rashba_str=0.02, KM_str=0.0, HLD_str=0.0, range=0.5;	
-
-	device_vars graphene_vars;
-	
-	//Reading device variables
-	graphene_vars.W_=256;
-	graphene_vars.LE_=256;
-	graphene_vars.C_=0.0;
-	graphene_vars.dis_str_=1.5;
-        graphene_vars.dis_seed_=1000;
-     	graphene_vars.SUBDIM_ = graphene_vars.W_*graphene_vars.LE_;
-	graphene_vars.DIM_    = graphene_vars.SUBDIM_ + 2 * graphene_vars.C_*graphene_vars.W_;
-
-
-
-	//Reading simulation parameters
-
-
-	  
-	Graphene_KaneMele *device = new Graphene_KaneMele (is_k_space, range, stgr_str, m_str,rashba_str,KM_str, HLD_str, graphene_vars);
-	device->build_Hamiltonian();
-
-
-	chebyshev::MomentsTD_Device chebMoms(*device, numMoms, numTimes); //load number of moments
-
-	
-	SparseMatrixType OP[2];
-	OP[0].SetID("HAM");
-	OP[1].SetID(S_OP);
-	
-	// Build the operators from Files
-	SparseMatrixBuilder builder;
-	std::array<double,2> spectral_bounds;	
-	for (int i = 0; i < 2; i++)
-	{
-		std::string input = "operators/" + LABEL + "." + OP[i].ID() + ".CSR";
-		builder.setSparseMatrix(&OP[i]);
-		builder.BuildOPFromCSRFile(input);
-	
-		if( i == 0 ) //is hamiltonian
-		//Obtain automatically the energy bounds
-		 spectral_bounds = chebyshev::utility::SpectralBounds(OP[0]);
-	};
-
-	spectral_bounds = chebyshev::utility::SpectralBounds(OP[0]);
-
-	
-	r_type a = (spectral_bounds[1]-spectral_bounds[0])*1.0, b = (spectral_bounds[1]+spectral_bounds[0])*0.5;
-
-	device->adimensionalize(a,b);  
-
-	//CONFIGURE THE CHEBYSHEV MOMENTS
-	chebMoms.SystemLabel(LABEL);
-	chebMoms.BandWidth ( (spectral_bounds[1]-spectral_bounds[0])*1.0);
-	chebMoms.BandCenter( (spectral_bounds[1]+spectral_bounds[0])*0.5);
-	chebMoms.TimeDiff( tmax/(numTimes-1) );
-	//chebMoms.SetAndRescaleHamiltonian(OP[0]);
-	chebMoms.Print();
-
-	
-
-	//Compute the chebyshev expansion table
-	qstates::generator gen;
-	if( argc == 7)  
-	  gen  = qstates::LoadStateFile(argv[6]);
-	
-
-	
-
-	device->Anderson_disorder();
-	chebyshev::TimeEvolvedOperator_Device( OP[1], chebMoms, gen);
-
-
-	
-	std::string outputfilename="Time"+S_OP+LABEL+"KPM_M"+S_NMOM+"_state"+gen.StateLabel()+".chebmomTD";	
-	std::cout<<"Saving convergence data in "<<outputfilename<<std::endl;
-	chebMoms.saveIn(outputfilename);
-	std::cout<<"End of program"<<std::endl;
-	return 0;
-}
-
-void time_evolution::printHelpMessage()
-	{
-		std::cout << "The program should be called with the following options: Label Op numMom numTimeSteps MaxTime (optional) num_states_file" << std::endl
-				  << std::endl;
-		std::cout << "Label will be used to look for Label.Ham, Label.Op " << std::endl;
-		std::cout << "Op will be used to located the sparse matrix file of two operators for the projection evolution" << std::endl;
-		std::cout << "numMom will be used to set the number of moments in the chebyshev table" << std::endl;
-		std::cout << "numTimeSteps  will be used to set the number of timesteps in the chebyshev table" << std::endl;
-		std::cout << "TimeMax  will be set the maximum time where the correlation will be evaluted " << std::endl;
-	};
-
-	inline
-void time_evolution::printWelcomeMessage()
-	{
-		std::cout << "WELCOME: This program will compute a table needed for expanding the correlation function in Chebyshev polynomialms" << std::endl;
-	};
